cleanup option handling.
cli.py now displays the description of the commands when given --help as the
only option. If --help is passed in as a param for the subcommand, those options
are now show correctly.

commands.py has a description parameter to the __init__ routine, and uses the
CLI options to override configuration options which in turn override application
defaults.

sm_photo_tool.py: fixed category handling in create_album